+++ rules/import.js
@@ -20,7 +20,10 @@ module.exports = {
   // Module systems
 
   // Style guide
-  "import/newline-after-import": "error",
-  "import/no-unassigned-import": ["error", { allow: ["**/*.css"] }],
-  "import/no-anonymous-default-export": "error",
+  "import/newline-after-import": "warn",
+  "import/no-unassigned-import": ["warn", { allow: ["**/*.css"] }],
+  "import/no-anonymous-default-export": [
+    "warn",
+    { allowCallExpression: false },
+  ],
 };